First version which solved Stefan-like problem!
It is not quite as it should be, there are no temperatures
solved here, but manages to push the fluid outside of the
domain.

Face fluxes are computed perfectly - which is really the
minimum we should have been hoping for, but velocities are
a bit ... well, they have a hunch across the boundary, and
I don't think it can be avoided if pressure is solved for
the entire water-steam system.

Also, plotting is much better than before - it plots cells
as well as contour of VOF.

new file:   plot_cells.m
new file:   plot_cells.m
modified:   plot_var.m
modified:   solver.par
modified:   tflows.m
modified:   vof.ini