IMPORTANT: BIG MISTAKE FIX!
For non-uniform grids, it was using the non-uniform
dx array, to compute dy and dz (which were thus non-
uniform themselves) which made the unique (scalar) sx
invalid and it also rendered discretization wrong.

Now both dy and dz are scalars, like sx, and are worked
out before the non-uniofmity is added in x direction.

modified:   tflows.m